requirements.txt
numpy==1.23.4
Pillow==9.2.0
PyMuPDF==1.20.2
